+++ bin/play.js
@@ -8,7 +8,7 @@ const chokidar = require('chokidar')
 const fileAccess = util.promisify(fs.access)
 const readFile = util.promisify(fs.readFile)
 const writeFile = util.promisify(fs.writeFile)
-const getSignature = require("./getSignature");
+const getSignature = require("../lib/getSignature")
 
 module.exports = async function createClient(options) {
   if (!options.url) {
