+++ lib/getProblemSet.js
@@ -4,7 +4,7 @@ const path = require('path')
 const chalk = require('chalk')
 const dedent = require('dedent')
 const readdir = util.promisify(fs.readdir)
-const { runTestFromFile } = require('./index')
+const { runTestFromFile } = require('./problemFarm')
 
 module.exports = async function getTests(dir) {
   console.log(`> building tests data from: ${chalk.cyan(dir)}`)
