+++ lib/getSignature.js
