+++ lib/problem.js
@@ -59,10 +59,7 @@ module.exports = async function runTest(test, solution, callback) {
     assert.strictEqual(typeof t.solution, 'function', `${id} must export solution`)
     assert.strictEqual(typeof t.description, 'string', `${id} must export description`)
     assert.strictEqual(typeof t.title, 'string', `${id} must export title`)
-
-    if (solution) {
-      assert.strictEqual(typeof s.solution, 'function', `${path.basename(solution.file)} must export solution`)
-    }
+    assert.strictEqual(typeof s.solution, 'function', `${path.basename(solution.file)} must export solution`)
 
     await t.test(s.solution)
